# Improve dict2imas & isequal